--- finish/backendServices/src/main/java/io/openliberty/guides/dao/EventDao.java
@@ -1,57 +1,0 @@
-// tag::copyright[]
-/*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - Initial implementation
- *******************************************************************************/
-// end::copyright[]
-package io.openliberty.guides.dao;
-
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
-import io.openliberty.guides.models.Event;
-
-import javax.enterprise.context.RequestScoped;
-import javax.transaction.Transactional;
-
-@RequestScoped
-@Transactional
-public class EventDao {
-
-    @PersistenceContext(name = "jpa-unit")
-    private EntityManager em;
-
-    public void createEvent(Event event) {
-        em.persist(event);
-    }
-
-    public Event readEvent(int eventId) {
-        return em.find(Event.class, eventId);
-    }
-
-    public void updateEvent(Event event) {
-        em.merge(event);
-    }
-
-    public void deleteEvent(Event event) {
-        em.remove(event);
-    }
-
-    public List<Event> readAllEvents() {
-        return em.createNamedQuery("Event.findAll", Event.class).getResultList();
-    }
-
-    public List<Event> findEvent(String name, String location, String time) {
-        return em.createNamedQuery("Event.findEvent", Event.class)
-            .setParameter("name", name)
-            .setParameter("location", location)
-            .setParameter("time", time).getResultList();
-    }
-}
