+++ finish/backendServices/src/test/java/it/io/openliberty/guides/event/EventEntityTest.java
@@ -25,7 +25,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import io.openliberty.guides.models.Event;
+import io.openliberty.guides.event.models.Event;
 
 public class EventEntityTest extends EventTest {
