+++ finish/frontendUI/src/main/webapp/META-INF/microprofile-config.properties
@@ -1,1 +1,1 @@
-io.openliberty.guides.client.EventClient/mp-rest/url=http://localhost:5050
+io.openliberty.guides.event.client.EventClient/mp-rest/url=http://localhost:5050
