--- start/frontendUI/src/main/java/io/openliberty/guides/client/BadRequestExceptionMapper.java
@@ -1,37 +1,0 @@
-// tag::copyright[]
-/*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - Initial implementation
- *******************************************************************************/
-// end::copyright[]
-package io.openliberty.guides.client;
-
-import java.util.logging.Logger;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.Provider;
-import org.eclipse.microprofile.rest.client.ext.ResponseExceptionMapper;
-import javax.ws.rs.BadRequestException;
-
-@Provider
-public class BadRequestExceptionMapper
-    implements ResponseExceptionMapper<BadRequestException> {
-  Logger LOG = Logger.getLogger(BadRequestExceptionMapper.class.getName());
-
-  @Override
-  public boolean handles(int status, MultivaluedMap<String, Object> headers) {
-    LOG.info("status = " + status);
-    return status == 400;
-  }
-
-  @Override
-  public BadRequestException toThrowable(Response response) {
-    return new BadRequestException(response.readEntity(String.class));
-  }
-}
