--- start/frontendUI/src/main/java/io/openliberty/guides/client/UnknownUrlExceptionMapper.java
@@ -1,36 +1,0 @@
-// tag::copyright[]
-/*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - Initial implementation
- *******************************************************************************/
-// end::copyright[]
-package io.openliberty.guides.client;
-
-import java.util.logging.Logger;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.Provider;
-import org.eclipse.microprofile.rest.client.ext.ResponseExceptionMapper;
-
-@Provider
-public class UnknownUrlExceptionMapper
-    implements ResponseExceptionMapper<UnknownUrlException> {
-  Logger LOG = Logger.getLogger(UnknownUrlExceptionMapper.class.getName());
-
-  @Override
-  public boolean handles(int status, MultivaluedMap<String, Object> headers) {
-    LOG.info("status = " + status);
-    return status == 404;
-  }
-
-  @Override
-  public UnknownUrlException toThrowable(Response response) {
-    return new UnknownUrlException();
-  }
-}
