--- start/frontendUI/src/main/java/io/openliberty/guides/ui/facelets/PageLoader.java
@@ -1,62 +1,0 @@
-// tag::copyright[]
-/*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - Initial implementation
- *******************************************************************************/
-// end::copyright[]
-package io.openliberty.guides.ui.facelets;
-
-import javax.inject.Named;
-import java.io.Serializable;
-import javax.enterprise.context.SessionScoped;
-
-@Named("pageLoader")
-@SessionScoped
-public class PageLoader implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private String content = "content/mainPage.xhtml";
-    private String headerContent = "header/header.xhtml";
-    private String navBar = "navBar/leftNav.xhtml";
-    private String currentPage = "Events";
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public String getNavBar() {
-        return navBar;
-    }
-
-    public void setNavBar(String navBar) {
-        this.navBar = navBar;
-    }
-
-    public String getHeaderContent() {
-        return this.headerContent;
-    }
-
-    public void setHeaderContent(String headerContent) {
-        this.headerContent = headerContent;
-    }
-
-    public String getCurrentPage() {
-        return this.currentPage;
-    }
-
-    public void setCurrentPage(String currentPage) {
-        this.currentPage = currentPage;
-    }
-
-}
